+++ services/container.py
@@ -25,6 +25,18 @@ from services.weather_system import WeatherSystem
 from services.time_system import TimeSystem
 from services.world_event_system import WorldEventSystem
 from services.quest_engine import QuestEngine
+from services.world_engine import WorldEngine
+from services.ecosystem_service import EcosystemService
+from services.artifact_service import ArtifactService
+from services.guild_territory import GuildTerritoryService
+from services.home_service import HomeService
+from services.npc_life_engine import NPCLifeEngine
+from services.world_memory_service import WorldMemoryService
+from services.guild_war_service import GuildWarService
+from services.npc_quest_service import NPCQuestService
+from services.seasonal_quest_service import SeasonalQuestService
+from services.world_chronicle_service import WorldChronicleService
+from services.world_boss_service import WorldBossService
 
 
 class ServiceContainer:
@@ -47,7 +59,6 @@ class ServiceContainer:
         self.crafting = CraftingService(self.chronicle, self.player, self.inventory)
         self.npc = NPCService(self.chronicle)
         self.npc_memory = NPCMemoryService(self.chronicle)
-        self.npc_scheduler = NPCScheduler(self.npc)
         self.exploration = ExplorationService(self.chronicle, self.player)
         self.catalog = CatalogService(self.chronicle)
         self.economy = EconomyService(self.chronicle, self.player)
@@ -57,7 +68,32 @@ class ServiceContainer:
         self.world_events = WorldEventSystem(self.chronicle)
         self.quest_engine = QuestEngine(self.chronicle)
 
-        self.user = self.player  # backward compat
+        self.ecosystem = EcosystemService(self.chronicle)
+        self.artifact = ArtifactService(self.chronicle)
+        self.guild_territory = GuildTerritoryService(self.chronicle)
+        self.home = HomeService(self.chronicle)
+        self.npc_life = NPCLifeEngine(self.chronicle)
+        self.world_memory = WorldMemoryService(self.chronicle)
+        self.guild_war = GuildWarService(self.chronicle)
+        self.npc_quest = NPCQuestService(self.chronicle)
+        self.seasonal_quest = SeasonalQuestService(self.chronicle)
+        self.world_chronicle = WorldChronicleService(self.chronicle)
+        self.world_boss = WorldBossService(self.chronicle)
+
+        self.npc_scheduler = NPCScheduler(self.npc)
+
+        self.world_engine = WorldEngine(
+            self.chronicle,
+            ecosystem=self.ecosystem,
+            guild_territory=self.guild_territory,
+            home_service=self.home,
+            npc_life=self.npc_life,
+            world_memory=self.world_memory,
+            seasonal_quest=self.seasonal_quest,
+            world_boss=self.world_boss,
+        )
+
+        self.user = self.player
 
 
 services = ServiceContainer()